feat(cli): add support for SW6 and fix params bug
Previously we'd naively generate the G1/G2 sizes by counting the size of the ToBytes representation of each element. This is wrong, as this would give us the 'max' size of the field. Counting the bytes in a serialized buffer produced by the CanonicalSerialize trait gives the desired sizes for the compressed elements. Uncompressed elements are just twice that size.